+++ src/MiniBlog/WebContent/WEB-INF/jsp/list-user.jsp
@@ -1,0 +1,55 @@
+<%@ page language="java" contentType="text/html; charset=ISO-8859-1"
+    pageEncoding="ISO-8859-1"%>
+<%@ include file="/WEB-INF/jsp/includes.jsp"%>
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<html>
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
+<title>User Management</title>
+</head>
+<body>
+<h1>Users Data</h1>
+<form:form action="user.do" method="POST" commandName="user">
+	<table>
+		<tr>
+			<td>User ID</td>
+			<td><form:input path="userID" /></td>
+		</tr>
+		<tr>
+			<td>User Name</td>
+			<td><form:input path="userName" /></td>
+		</tr>
+		<tr>
+			<td>First name</td>
+			<td><form:input path="firstName" /></td>
+		</tr>
+		<tr>
+			<td>Last name</td>
+			<td><form:input path="lastName" /></td>
+		</tr>
+		
+		<tr>
+			<td colspan="2">
+				<input type="submit" name="action" value="Add" />
+				<input type="submit" name="action" value="Edit" />
+				<input type="submit" name="action" value="Delete" />
+				<input type="submit" name="action" value="Search" />
+			</td>
+		</tr>
+	</table>
+</form:form>
+<br>
+<table border="1">
+	<th>ID</th>
+	<th>First name</th>
+	<th>Last name</th>
+	<c:forEach items="${userList}" var="user">
+		<tr>
+			<td>${user.userID}</td>
+			<td>${user.firstName}</td>
+			<td>${user.lastName}</td>
+		</tr>
+	</c:forEach>
+</table>
+</body>
+</html>
